+++ CHANGELOG.md
@@ -26,6 +26,7 @@ The following changes have been completed, and will be included in the next tagg
 ### New
 
 * Added `isMediaType()` data guard
+* Added MediaTypeParts value type
 * Added MediaTypeMatchRegex
 * Added MediaTypeParamRegex